+++ src/api/index.js
@@ -1,5 +1,6 @@
 // @flow
 
 import connectChangelog from './connectChangelog';
+import { CURRENT_VERSION } from './migrations';
 
-export { connectChangelog };
+export { connectChangelog, CURRENT_VERSION };